POC: Implement login components so that import component based on build target
